Updated RTL front end unit tests and removed Enzyme tests
