+++ src/app/__tests__/Loader.test.tsx
@@ -10,13 +10,11 @@ describe('unit testing for Loader.tsx', () => {
   });
 
   test('renders a fail icon', () => {
-    // render(<Loader loading={false} result={false} />);
     const { container } = render(<Loader loading={false} result={false} />);
     expect(container.getElementsByClassName('fail').length).toBe(1);
   });
 
   test('renders a check icon', () => {
-    // render(<Loader loading={false} result={true} />);
     const { container } = render(<Loader loading={false} result={true} />);
     expect(container.getElementsByClassName('check').length).toBe(1);
   });
